feat(frontend): rework
